requirements.txt
mcp>=0.1.0
resemble>=1.0.0
python-dotenv>=0.19.0